Move Ant, Maven, Cli to separate packages
